Deprecate core video utilities and refactor to improve module organization:
- Remove unused video-related utilities from `core/video.py`.
- Relocate and reorganize rate limiter, pagination, authentication, and RabbitMQ subscriptions within appropriate `src` modules.
- Export key core functions in `core/__init__.py` for better modularity and maintainability.
