+++ backend/src/core/__init__.py
@@ -1,0 +1,9 @@
+from .background_tasks import deindex_video_in_es, index_video_in_es, update_video_in_es
+from .pagination import paginate_query
+
+__all__ = [
+    "paginate_query",
+    "index_video_in_es",
+    "deindex_video_in_es",
+    "update_video_in_es",
+]
